+++ pwsp-gui/src/gui/views/footer.rs
@@ -1,5 +1,5 @@
 use crate::gui::SoundpadGui;
-use egui::{AtomExt, Button, ComboBox, Label, RichText, Slider, Ui, Vec2};
+use egui::{AtomExt, Button, Label, RichText, Slider, Ui, Vec2};
 use egui_material_icons::icons::*;
 use pwsp_lib::types::gui::SliderLatch;
 use rust_i18n::t;
@@ -7,7 +7,8 @@ use rust_i18n::t;
 /// Masters go past 100% on purpose: amplifying the microphone feed without deafening
 /// yourself is the reason the two paths were split in the first place.
 const MAX_MASTER_VOLUME: f32 = 2.0;
-const VOLUME_SLIDER_WIDTH: f32 = 110.0;
+const VOLUME_SLIDER_WIDTH: f32 = 90.0;
+const ICON_SIZE: f32 = 18.0;
 
 impl SoundpadGui {
     pub fn draw_footer(&mut self, ui: &mut Ui) {
@@ -16,69 +17,15 @@ impl SoundpadGui {
             self.draw_monitoring_volume(ui);
             self.draw_mic_volume(ui);
 
-            ui.add_space(ui.available_width() - 18.0 * 2.0 - ui.spacing().item_spacing.x * 2.0);
+            // Pushes the two icon buttons to the right edge. Clamped because a window
+            // narrow enough to make this negative would otherwise shove them out of
+            // view instead of just crowding them.
+            let spacer = ui.available_width() - ICON_SIZE * 2.0 - ui.spacing().item_spacing.x * 2.0;
+            ui.add_space(spacer.max(0.0));
 
             self.draw_hotkeys_button(ui);
             self.draw_settings_button(ui);
         });
-        ui.add_space(5.0);
-        ui.horizontal(|ui| {
-            self.draw_mic_selection(ui);
-            self.draw_output_selection(ui);
-        });
-    }
-
-    fn draw_mic_selection(&mut self, ui: &mut Ui) {
-        let mics = &self.audio_player_state.all_inputs_sorted;
-
-        let mut selected_input = self.audio_player_state.current_input.to_owned();
-        let prev_input = selected_input.to_owned();
-        ComboBox::from_label(t!("gui.choose_mic_select"))
-            .height(30.0)
-            .selected_text(
-                self.audio_player_state
-                    .all_inputs
-                    .get(&selected_input)
-                    .unwrap_or(&String::new()),
-            )
-            .show_ui(ui, |ui| {
-                for (name, nick) in mics {
-                    ui.selectable_value(&mut selected_input, name.clone(), nick);
-                }
-            });
-
-        if selected_input != prev_input {
-            self.set_input(selected_input);
-        }
-    }
-
-    fn draw_output_selection(&mut self, ui: &mut Ui) {
-        let outputs = &self.audio_player_state.all_outputs_sorted;
-
-        let mut selected_output = self.audio_player_state.current_output.to_owned();
-        let prev_output = selected_output.to_owned();
-
-        // An empty selection means no device is pinned and playback follows the system
-        // default sink, which is also what a fresh install does.
-        let selected_text = self
-            .audio_player_state
-            .all_outputs
-            .get(&selected_output)
-            .cloned()
-            .unwrap_or_else(|| t!("gui.default_output").to_string());
-
-        ComboBox::from_label(t!("gui.choose_output_select"))
-            .height(30.0)
-            .selected_text(selected_text)
-            .show_ui(ui, |ui| {
-                for (name, nick) in outputs {
-                    ui.selectable_value(&mut selected_output, name.clone(), nick);
-                }
-            });
-
-        if selected_output != prev_output {
-            self.set_output(selected_output);
-        }
     }
 
     fn draw_monitoring_volume(&mut self, ui: &mut Ui) {
@@ -110,11 +57,16 @@ impl SoundpadGui {
     fn draw_volume_slider(ui: &mut Ui, latch: &mut SliderLatch, daemon_value: f32) {
         latch.sync(daemon_value);
 
+        // A Slider draws its rail at spacing().slider_width no matter what add_sized
+        // allocates, so both have to be told the same number or the widget overruns the
+        // space reserved for it.
+        ui.spacing_mut().slider_width = VOLUME_SLIDER_WIDTH;
+
         let slider = Slider::new(&mut latch.value, 0.0..=MAX_MASTER_VOLUME)
             .show_value(false)
             .step_by(0.01);
         if ui
-            .add_sized([VOLUME_SLIDER_WIDTH, 18.0], slider)
+            .add_sized([VOLUME_SLIDER_WIDTH, ICON_SIZE], slider)
             .drag_stopped()
         {
             latch.dragged = true;
